+++ aiomoqt/examples/multi_sub_bench.py
@@ -21,6 +21,7 @@ import sys
 import time
 
 from aiomoqt.client import MOQTClient
+from aiomoqt.types import parse_draft_spec
 from aiomoqt.track import PublishedTrack, VideoTrack, SubscribedTrack
 from aiomoqt.utils import wait_cond_timeout
 from aiomoqt.utils.logger import set_log_level
@@ -61,7 +62,7 @@ def parse_args():
                         help='Duration seconds (default: 30)')
     parser.add_argument('-k', '--insecure', action='store_true',
                         help='Skip TLS verification')
-    parser.add_argument('--draft', type=int, default=None,
+    parser.add_argument('--draft', type=parse_draft_spec, default=None,
                         help='MoQT draft version (default: negotiate)')
     parser.add_argument('--video', type=str, default=None,
                         choices=['240p', '270p', '360p', '480p',